requirements.txt
absl-py==1.0.0
amqp==5.0.5
APScheduler==3.8.1
asgiref==3.3.1
astroid==2.5.2
astunparse==1.6.3
backports.entry-points-selectable==1.1.0
billiard==3.6.3.0
boto3==1.17.11
botocore==1.20.11
cachetools==4.2.4
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.1.7
django-allauth==0.51.0
django-anymail==8.4
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-ses==1.0.3
djangorestframework==3.12.2
djangorestframework-filters==1.0.0.dev0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==4.8.0
docopt==0.6.2
filelock==3.1.0
flatbuffers==2.0
func-timeout==4.3.5
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==2.10
importlib-metadata==4.8.2
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.6.0
libclang==12.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.1
openapi-codec==1.3.2
opt-einsum==3.3.0
pipreqs==0.4.10
platformdirs==2.4.0
prompt-toolkit==3.0.16
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pylint==2.7.4
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.1
python3-openid==3.2.0
pytweening==1.0.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.3.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
yarg==0.1.9
